Create Random Selection of a Person to Cover the Restaurant Bill.
The provided Python program allows users to input a list of names separated by commas and spaces. It then randomly selects one of the names and announces that the selected person will be responsible for covering the restaurant bill. This simple program demonstrates the concept of random selection and can be used for fun or lighthearted decision-making in a group setting.